adding tests in the plugin
