+++ buildSrc/src/main/kotlin/com/talaiot/buildplugins/JacocoExtensions.kt
@@ -7,15 +7,17 @@ import org.gradle.testing.jacoco.plugins.JacocoPluginExtension
 import java.io.File
 
 fun Project.setUpJacoco() = this.run {
+    configure<JacocoPluginExtension> {
+        toolVersion = "0.8.6"
+    }
+
     tasks.withType<JacocoReport> {
         reports {
             xml.isEnabled = true
-            csv.isEnabled = true
+            csv.isEnabled = false
             html.isEnabled = true
             html.destination = File("${rootProject.buildDir}/coverage")
         }
     }
-    configure<JacocoPluginExtension> {
-        toolVersion = "0.8.3"
-    }
+
 }
