+++ buildSrc/src/main/kotlin/com/talaiot/buildplugins/TalaiotPlugin.kt
@@ -1,10 +1,8 @@
 package com.talaiot.buildplugins
 
-import com.gradle.publish.PluginBundleExtension
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.*
-import org.gradle.plugin.devel.GradlePluginDevelopmentExtension
 
 /**
  * Talaiot Plugin abstracts the build logic for modules used as Gradle Plugin.
@@ -13,7 +11,7 @@ import org.gradle.plugin.devel.GradlePluginDevelopmentExtension
 class TalaiotPlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
-        val extension = target
+        target
             .extensions
             .create<TalaiotPluginConfiguration>("talaiotPlugin")
 
@@ -21,16 +19,13 @@ class TalaiotPlugin : Plugin<Project> {
         target.plugins.apply("maven-publish")
         target.plugins.apply("jacoco")
         target.plugins.apply("kotlin")
+        target.plugins.apply("java-library")
+        target.plugins.apply("application")
         target.plugins.apply("com.gradle.plugin-publish")
 
         target.repositories {
             jcenter()
             mavenCentral()
-        }
-
-        target.dependencies {
-            add("testImplementation", "com.nhaarman.mockitokotlin2:mockito-kotlin:2.0.0-RC1")
-            add("testImplementation", "io.kotlintest:kotlintest-runner-junit5:3.3.2")
         }
 
         target.setUpJacoco()
@@ -42,33 +37,14 @@ class TalaiotPlugin : Plugin<Project> {
             setProjectGroup(extension.group, Constants.DEFAULT_GROUP_PLUGIN)
             collectUnitTest()
             setUpPublishing("mavenTalaiot", extension.artifact)
-            configGradlePlugin(target)
-        }
-    }
-
-    private fun configGradlePlugin(project: Project) {
-
-        project.configure<GradlePluginDevelopmentExtension> {
-            plugins {
-                register(project.name) {
-                    val extension = project.extensions.getByType<TalaiotPluginConfiguration>()
-                    id = extension.idPlugin
-                    implementationClass = extension.mainClass
-                }
-            }
+            setUpGradlePublishing()
         }
 
-        project.configure<PluginBundleExtension>() {
-            (plugins){
-                (project.name){
-                    val extension = project.extensions.getByType<TalaiotPluginConfiguration>()
-                    displayName = project.name
-                    description =
-                        "Simple and extensible plugin to track task and build times in your Gradle Project."
-                    tags = listOf("tracking", "kotlin", "gradle")
-                    version = Constants.TALAIOT_VERSION
-                }
-            }
+        target.dependencies {
+            add("testImplementation", "com.nhaarman.mockitokotlin2:mockito-kotlin:2.0.0-RC1")
+            add("testImplementation", "io.kotlintest:kotlintest-runner-junit5:3.3.2")
         }
+
     }
 }
+
