+++ library/talaiot-test-utils/src/main/java/com/cdsap/talaiot/logger/TestLogTrackerRecorder.kt
@@ -1,4 +1,4 @@
-package com.cdsap.talaiot.utils
+package com.cdsap.talaiot.logger
 
 import com.cdsap.talaiot.logger.LogTracker
