+++ library/talaiot-test-utils/src/main/java/org/testcontainers/elasticsearch/KElasticSearchContainer.kt
@@ -1,4 +1,4 @@
-package com.cdsap.talaiot.publisher
+package org.testcontainers.elasticsearch
 
 import org.testcontainers.elasticsearch.ElasticsearchContainer
