+++ library/talaiot-test-utils/src/main/java/org/testcontainers/influxdb/KInfluxDBContainer.kt
@@ -1,4 +1,4 @@
-package com.cdsap.talaiot.publisher.graphpublisher
+package org.testcontainers.influxdb
 
 import org.testcontainers.containers.InfluxDBContainer
