+++ library/talaiot-test-utils/src/main/java/org/testcontainers/remotecache/KRedisContainer.kt
@@ -1,4 +1,4 @@
-package com.cdsap.talaiot.publisher
+package org.testcontainers.remotecache
 
 import org.testcontainers.remotecache.RedisRemoteCacheContainer
