+++ library/talaiot/src/test/kotlin/com/cdsap/talaiot/publisher/InfluxDbPublisherTest.kt
@@ -1,9 +1,10 @@
 package com.cdsap.talaiot.publisher
 
 import com.cdsap.talaiot.configuration.InfluxDbPublisherConfiguration
-import com.cdsap.talaiot.logger.TestLogTrackerRecorder
-import com.cdsap.talaiot.publisher.graphpublisher.KInfluxDBContainer
+import org.testcontainers.influxdb.KInfluxDBContainer
 import com.cdsap.talaiot.report.ExecutionReportProvider
+import com.cdsap.talaiot.utils.TestExecutor
+import com.cdsap.talaiot.logger.TestLogTrackerRecorder
 import io.kotlintest.Spec
 import io.kotlintest.specs.BehaviorSpec
 import org.influxdb.dto.Query
