+++ library/talaiot/src/test/kotlin/com/cdsap/talaiot/publisher/PushGatewayPublisherTest.kt
@@ -3,15 +3,17 @@ package com.cdsap.talaiot.publisher
 import com.cdsap.talaiot.configuration.PushGatewayPublisherConfiguration
 import com.cdsap.talaiot.entities.*
 
-import com.cdsap.talaiot.logger.TestLogTrackerRecorder
 import com.cdsap.talaiot.publisher.pushgateway.PushGatewayFormatter
 import com.cdsap.talaiot.publisher.pushgateway.PushGatewayPublisher
 import com.cdsap.talaiot.report.ExecutionReportProvider
 import com.cdsap.talaiot.request.SimpleRequest
+import com.cdsap.talaiot.utils.TestExecutor
+import com.cdsap.talaiot.logger.TestLogTrackerRecorder
 import io.github.rybalkinsd.kohttp.dsl.httpGet
 import io.github.rybalkinsd.kohttp.ext.url
 import io.kotlintest.Spec
 import io.kotlintest.specs.BehaviorSpec
+import org.testcontainers.pushgateway.KPushGatewayContainer
 import java.net.URL
