+++ library/talaiot/src/test/kotlin/com/cdsap/talaiot/publisher/RethinkDbPublisherTest.kt
@@ -6,6 +6,7 @@ import com.cdsap.talaiot.entities.*
 import com.cdsap.talaiot.logger.TestLogTrackerRecorder
 import com.cdsap.talaiot.publisher.rethinkdb.RethinkDbPublisher
 import com.cdsap.talaiot.report.ExecutionReportProvider
+import com.cdsap.talaiot.utils.TestExecutor
 import com.rethinkdb.RethinkDB
 import com.rethinkdb.net.Connection
 import com.rethinkdb.net.Cursor
@@ -15,6 +16,7 @@ import io.kotlintest.inspectors.forAtLeastOne
 import io.kotlintest.shouldThrow
 import io.kotlintest.specs.BehaviorSpec
 import junit.framework.Assert.assertTrue
+import org.testcontainers.rethinkdb.KRethinkDbContainer
 import java.net.URL
 
 class RethinkDbPublisherTest : BehaviorSpec() {
