+++ library/talaiot/src/test/kotlin/com/cdsap/talaiot/publisher/graphpublisher/DotPublisherTest.kt
@@ -2,7 +2,7 @@ package com.cdsap.talaiot.publisher.graphpublisher
 
 import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.mock.TaskMeasurementAggregatedMock
-import com.cdsap.talaiot.publisher.TestExecutor
+import com.cdsap.talaiot.utils.TestExecutor
 import com.cdsap.talaiot.writer.FileWriter
 import com.nhaarman.mockitokotlin2.*
 import io.kotlintest.specs.BehaviorSpec
