+++ library/talaiot/src/test/kotlin/com/cdsap/talaiot/publisher/graphpublisher/HtmlPublisherTest.kt
@@ -3,8 +3,8 @@ package com.cdsap.talaiot.publisher.graphpublisher
 import com.cdsap.talaiot.publisher.graphpublisher.resources.ResourcesHtml
 import com.cdsap.talaiot.logger.LogTracker
 import com.cdsap.talaiot.mock.TaskMeasurementAggregatedMock.taskMeasurementAggregated
-import com.cdsap.talaiot.publisher.TestExecutor
 import com.cdsap.talaiot.publisher.graphpublisher.resources.ResourcesHtml.LEGEND_HEADER
+import com.cdsap.talaiot.utils.TestExecutor
 import com.cdsap.talaiot.writer.FileWriter
 import com.nhaarman.mockitokotlin2.argThat
 import com.nhaarman.mockitokotlin2.mock
